+++ orbit/models/ktr.py
@@ -29,7 +29,8 @@ from ..estimators.stan_estimator import StanEstimatorMAP, StanEstimatorMCMC
 from ..exceptions import IllegalArgument, ModelException, PredictionException
 from .base_model import BaseModel
 from ..utils.general import is_ordered_datetime
-from ..utils.kernels import gauss_kernel, sandwich_kernel, generate_seas
+from ..utils.kernels import gauss_kernel, sandwich_kernel
+from ..utils.timepoints import generate_seas
 from ..utils.features import make_fourier_series_df, make_fourier_series
